rename stage_init.py -> stage_init_utils.py and align comments with refactor branch
Resolves merge conflicts with fake0fan/refactor:
- Rename stage_init.py to stage_init_utils.py to match refactor branch
- Update all imports to reference the new module name
- Align comment style (remove V0 references) with refactor branch

Made-with: Cursor
